update :: delete model universal
